refactor : MVC로 나눔
