+++ src/main/java/baseball/domain/Game.java
@@ -1,20 +1,26 @@
 package baseball.domain;
 
+import java.util.ArrayList;
+import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Set;
+
+import baseball.common.Number;
+import nextstep.utils.Randoms;
 
 public class Game {
+	
+	private List<String> targetNumbers;
 	
 	private int strike;
 	private int ball;
 	
 	public Game() {
-		this.strike = 0;
-		this.ball = 0;
-	}
-
-	public Game(int strike, int ball) {
-		this.strike = strike;
-		this.ball = ball;
+		Set<String> numbers = new LinkedHashSet<String>();
+		while (numbers.size() < Number.TOTAL_TARGET_NUMBER) {
+			numbers.add(Integer.toString(Randoms.pickNumberInRange(Number.RANDOM_MIN_NUMBER, Number.RANDOM_MAX_NUMBER)));
+		}
+		this.targetNumbers = new ArrayList<>(numbers);
 	}
 
 	public int getStrike() {
@@ -25,18 +31,31 @@ public class Game {
 		return ball;
 	}
 	
-	public Game countBall(List<String> target, String number) {
+	private void countBall(List<String> target, String number) {
 		if (target.contains(number)) {
 			ball++;
 		}
-		return this;
 	}
 	
-	public Game countStrike(String target, String number) {
+	private void countStrike(String target, String number) {
 		if (target.equals(number)) {
 			strike++;
 			ball--;
 		}
-		return this;
+	}
+	
+	public void resetHint() {
+		this.strike = 0;
+		this.ball = 0;
+	}
+	
+	public void countAnswer(String playerNumber) {
+		resetHint();
+		String[] playNumbers = playerNumber.split("");
+		for (int i = 0; i < playNumbers.length; i++) {
+			countBall(targetNumbers, playNumbers[i]);
+			countStrike(targetNumbers.get(i), playNumbers[i]);
+		}
+		
 	}
 }
